requirements.txt
Django==3.0.5
django-import-export==2.0.2
djangorestframework==3.10.3
django-filter==2.2.0